add capella/deneb changes
